Codacy: HabitRecord: fixing parameterobject
