+++ app/src/main/java/com/ivanmagda/habito/models/HabitRecord.java
@@ -41,18 +41,19 @@ public final class HabitRecord implements Parcelable {
         this.checkmarks = new ArrayList<>();
     }
 
-    public HabitRecord(HabitRecordParameterObject paramObject) {
-        this.userId = paramObject.getUserId();
-        this.createdAt = paramObject.getCreatedAt();
-        this.name = paramObject.getName();
-        this.color = paramObject.getColor();
-        this.target = paramObject.getTarget();
-        this.resetFreq = paramObject.getResetFreq();
-        this.resetTimestamp = paramObject.getResetTimestamp();
-        this.reminderHour = paramObject.getReminderHour();
-        this.reminderMin = paramObject.getReminderMin();
-        this.score = paramObject.getScore();
-        this.checkmarks = paramObject.getCheckmarks();
+    public HabitRecord(String userId, long createdAt, String name, int color, int target,
+                       ResetParameterObject resetParam, ReminderParameterObject reminderParam, int score, List<Long> checkmarks) {
+        this.userId = userId;
+        this.createdAt = createdAt;
+        this.name = name;
+        this.color = color;
+        this.target = target;
+        this.resetFreq = resetParam.getResetFreq();
+        this.resetTimestamp = resetParam.getResetTimestamp();
+        this.reminderHour = reminderParam.getReminderHour();
+        this.reminderMin = reminderParam.getReminderMin();
+        this.score = score;
+        this.checkmarks = checkmarks;
     }
 
     public HabitRecord(Parcel in) {
@@ -103,8 +104,9 @@ public final class HabitRecord implements Parcelable {
     };
 
     public HabitRecord copy() {
-        return new HabitRecord(
-                new HabitRecordParameterObject(userId, createdAt, name, color, target, resetFreq, resetTimestamp, reminderHour, reminderMin, score, checkmarks));
+        return new HabitRecord(userId, createdAt, name, color, target,
+                new ResetParameterObject(resetFreq, resetTimestamp),
+                new ReminderParameterObject(reminderHour, reminderMin), score, checkmarks);
     }
 
     public String getUserId() {
