+++ app/src/main/java/com/ivanmagda/habito/models/ResetParameterObject.java
@@ -1,0 +1,19 @@
+package com.ivanmagda.habito.models;
+
+public class ResetParameterObject {
+    private final String resetFreq;
+    private final long resetTimestamp;
+
+    public ResetParameterObject(String resetFreq, long resetTimestamp) {
+        this.resetFreq = resetFreq;
+        this.resetTimestamp = resetTimestamp;
+    }
+
+    public String getResetFreq() {
+        return resetFreq;
+    }
+
+    public long getResetTimestamp() {
+        return resetTimestamp;
+    }
+}
